+++ src/Beef.AspNetCore.WebApi/WebApiExtensions.cs
@@ -12,19 +12,24 @@ namespace Beef.AspNetCore.WebApi
     public static class WebApiExtensions
     {
         /// <summary>
-        /// Registers the action to <paramref name="update"/> the <see cref="ExecutionContext"/> for a request. 
+        /// Registers the action to <paramref name="updateAction"/> the <see cref="ExecutionContext"/> for a request. 
         /// </summary>
         /// <param name="builder">The <see cref="IApplicationBuilder"/>.</param>
-        /// <param name="update">The function to update the <see cref="ExecutionContext"/> for a request; <c>null</c> will <see cref="ExecutionContext.Reset(bool)"/> the <b>ExecutionContext</b>.</param>
+        /// <param name="updateAction">The optional <see cref="Action{HttpContext, ExecutionContext}"/>.</param>
         /// <returns>The <see cref="IApplicationBuilder"/>.</returns>
-        public static IApplicationBuilder UseExecutionContext(this IApplicationBuilder builder, Action<HttpContext, ExecutionContext> update = null)
+        public static IApplicationBuilder UseExecutionContext(this IApplicationBuilder builder, Action<HttpContext, ExecutionContext> updateAction = null)
         {
-            return builder.Use((context, next) =>
-            {
-                ExecutionContext.Reset(true);
-                update?.Invoke(context, ExecutionContext.Current);
-                return next();
-            });
+            return builder.UseMiddleware<WebApiExecutionContextMiddleware>(updateAction ?? WebApiExecutionContextMiddleware.DefaultUpdateAction);
+        }
+
+        /// <summary>
+        /// Adds <see cref="WebApiExceptionHandlerMiddleware"/> to the <see cref="IApplicationBuilder"/> request execution pipeline.
+        /// </summary>
+        /// <param name="builder">The <see cref="IApplicationBuilder"/>.</param>
+        /// <returns>The <see cref="IApplicationBuilder"/>.</returns>
+        public static IApplicationBuilder UseWebApiExceptionHandler(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<WebApiExceptionHandlerMiddleware>();
         }
     }
 }