+++ templates/Beef.Template.Solution/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Represents the **NuGet** versions.
 
+## v2.1.6
+- *Fixed:* Updated referenced *Beef* NuGet references to latest.
+
 ## v2.1.5
 - *Fixed:* Updated referenced *Beef* NuGet references to latest.
 - *Fixed:* `ExpectValidationException.Run` renamed to `ExpectValidationException.Throws` to correct example Test code.